Speedup the MarketHistory by using doubles.

+++ CryptoTrader/Trading/MarketAnalysis/MarketAggregatedHistory.swift
@@ -7,7 +7,7 @@ final class MarketAggregatedHistory: MarketHistorySlice {
     private var lastCleanup = 0
     
     private var currentTrade: DatedPrice? = nil
-    private var currentTradeAggregationCount: Decimal = 0.0
+    private var currentTradeAggregationCount: Double = 0.0
     private var aggregationPeriod: TimeInterval
     
     init(intervalToKeep: TimeInterval, aggregationPeriod: TimeInterval = 1) {
@@ -40,6 +40,7 @@ final class MarketAggregatedHistory: MarketHistorySlice {
             }
         }
         else {
+                        
             let secondAggregatedTrade =
                 DatedPrice(
                     price: (currentTrade.price * currentTradeAggregationCount + newTrade.price) / (currentTradeAggregationCount + 1),
